requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
colorama>=0.4.6  # Untuk warna di terminal
requests>=2.31.0
